+++ internal/config/config.go
@@ -30,9 +30,9 @@ var (
 	//
 	// DO NOT FORGET TO MODIFY VERSIONS IN BaseOperatorConf
 	defaultEnvs = map[string]string{
-		"VM_METRICS_VERSION": "v1.119.0",
-		"VM_LOGS_VERSION":    "v1.23.3",
-		"VM_ANOMALY_VERSION": "v1.23.2",
+		"VM_METRICS_VERSION": "v1.120.0",
+		"VM_LOGS_VERSION":    "v1.24.0",
+		"VM_ANOMALY_VERSION": "v1.24.0",
 	}
 )
 
@@ -98,8 +98,8 @@ type BaseOperatorConf struct {
 	//
 	// DO NOT FORGET TO MODIFY VERSIONS IN defaultEnvs
 
-	MetricsVersion string `default:"v1.119.0" env:"METRICS_VERSION"`
-	LogsVersion    string `default:"v1.23.3" env:"LOGS_VERSION"`
+	MetricsVersion string `default:"v1.120.0" env:"METRICS_VERSION"`
+	LogsVersion    string `default:"v1.24.0" env:"LOGS_VERSION"`
 	AnomalyVersion string `default:"v1.24.0" env:"ANOMALY_VERSION"`
 
 	// enables custom config reloader for vmauth and vmagent,